+++ src/camera/camera2d.jl
@@ -5,6 +5,7 @@ struct Camera2D <: AbstractCamera
     panbutton::Node{ButtonTypes}
     padding::Node{Float32}
     last_area::Node{Vec{2, Int}}
+    update_limits::Node{Bool}
 end
 
 """
@@ -13,15 +14,16 @@ end
 Creates a 2D camera for the given Scene.
 """
 function cam2d!(scene::SceneLike; kw_args...)
-    cam_attributes, rest = merged_get!(:cam2d, scene, Attributes(kw_args)) do
+    cam_attributes = merged_get!(:cam2d, scene, Attributes(kw_args)) do
         Theme(
             area = node(:area, FRect(0, 0, 1, 1)),
             zoomspeed = 0.10f0,
             zoombutton = nothing,
-            panbutton = Mouse.right,
+            panbutton = Mouse.left,
             selectionbutton = (Keyboard.space, Mouse.left),
             padding = 0.001,
-            last_area = Vec(size(scene))
+            last_area = Vec(size(scene)),
+            update_limits = false,
         )
     end
     cam = from_dict(Camera2D, cam_attributes)
@@ -86,6 +88,11 @@ function update_cam!(scene::SceneLike, cam::Camera2D)
     camera(scene).projection[] = projection
     camera(scene).projectionview[] = projection * view
     cam.last_area[] = Vec(size(scene))
+    if cam.update_limits[]
+        #
+        w2 = Vec2f0(w, h) .* 0.2
+        update_limits!(scene, Rect(origin(cam.area[]) .+ w2, widths(cam.area[]) .- 2w2))
+    end
     return
 end
 
@@ -165,11 +172,16 @@ function absrect(rect)
     end
     FRect(Vec2f0(xy), Vec2f0(abs.(wh)))
 end
+
+
 function selection_rect!(scene, cam, key)
     rect = RefValue(FRect())
     lw = 2f0
-    scene_unscaled = Scene(scene, transformation = Transformation(), cam = copy(camera(scene)))
-    theme(scene_unscaled, :clear)[] = false
+    scene_unscaled = Scene(
+        scene, transformation = Transformation(),
+        cam = copy(camera(scene)), clear = false
+    )
+    scene_unscaled.clear = false
     scene_unscaled.updated = Node(false)
     rect_vis = lines!(
         scene_unscaled,
@@ -215,9 +227,6 @@ function selection_rect!(scene, cam, key)
     end
     rect_vis, dragged_rect
 end
-
-
-
 
 function reset!(cam, boundingbox, preserveratio = true)
     w1 = widths(boundingbox)