+++ src/interaction/gui.jl
@@ -263,10 +263,21 @@ struct Popup
 end
 
 
-function textslider(range, label, scene = Scene(camera = campixel!); start = first(range), textalign = (:left, :center), textpos = (0, 50), kwargs...)
-    t = text!(scene, "$label:", raw = true, position = textpos, align = textalign, kwargs...)[end]
+function textslider(
+        range, label, scene = Scene(camera = campixel!);
+        start = first(range), textalign = (:left, :center), textpos = (0, 50),
+        textcolor = :black,
+        kwargs...
+    )
+    t = text!(
+        scene, "$label:", raw = true, position = textpos, align = textalign,
+        color = textcolor, kwargs...
+    )[end]
     xp = widths(boundingbox(t))[1]
-    s = slider!(scene, range, position = Point2f0(xp, 0), raw = true, start = start, kwargs...)[end]
+    s = slider!(
+        scene, range, position = Point2f0(xp, 0), raw = true,
+        start = start, textcolor = textcolor, kwargs...
+    )[end]
     scene, s[:value]
 end
 
@@ -277,8 +288,7 @@ function sample_color(f, ui, colormesh, v)
         ui, transformation = Transformation(), px_area = pixelarea(ui),
         theme = theme(ui)
     )
-    # FIXME, we need an api for this!
-    sub.theme.attributes[:clear] = Observable(false)
+    sub.clear = false
     select = scatter!(
         sub, lift((p, a)-> [Point2f0(p) .- minimum(a)], mpos, pixelarea(sub)),
         markersize = 15, color = (:white, 0.2), strokecolor = :white,
@@ -300,8 +310,6 @@ function sample_color(f, ui, colormesh, v)
         return
     end
 end
-
-
 
 function popup(parent, position, width)
     pos_n = Node(Point2f0(position))
@@ -336,7 +344,7 @@ function popup(parent, position, width)
 end
 
 """
-    Colorswatch
+    colorswatch(scene = Scene(camera = campixel!))
 
 TODO add function signatures
 TODO add description