+++ src/interfaces.jl
@@ -89,11 +89,12 @@ $(ATTRIBUTES)
         absorption = 1f0,
         isovalue = 0.5f0,
         isorange = 0.05f0,
+        color = nothing,
         colormap = theme(scene, :colormap),
         colorrange = (0, 1),
-        color = nothing,
     )
 end
+
 function mutual_exclusive_attributes(::Type{<:Volume})
     Dict(
         :colorrange => :color,
@@ -242,9 +243,9 @@ $(ATTRIBUTES)
 @recipe(Text, text) do scene
     Theme(;
         default_theme(scene)...,
+        font = theme(scene, :font),
         strokecolor = (:black, 0.0),
         strokewidth = 0,
-        font = theme(scene, :font),
         align = (:left, :bottom),
         rotation = 0.0,
         textsize = 20,
@@ -253,8 +254,8 @@ $(ATTRIBUTES)
 end
 
 const atomic_function_symbols = (
-        :text, :meshscatter, :scatter, :mesh, :linesegments,
-        :lines, :surface, :volume, :heatmap, :image
+    :text, :meshscatter, :scatter, :mesh, :linesegments,
+    :lines, :surface, :volume, :heatmap, :image
 )
 
 
@@ -312,12 +313,13 @@ function calculated_attributes!(::Type{<: MeshScatter}, plot)
     color_and_colormap!(plot)
 end
 
+
 function calculated_attributes!(::Type{<: Scatter}, plot)
     # calculate base case
     color_and_colormap!(plot)
     replace_automatic!(plot, :marker_offset) do
         # default to middle
-        lift(x-> Vec2f0.((x .* (-0.5f0))), plot[:markersize])
+        lift(x-> to_2d_scale(x .* (-0.5f0)), plot[:markersize])
     end
 end
 
@@ -424,16 +426,12 @@ function (PlotType::Type{<: AbstractPlot{Typ}})(scene::SceneLike, attributes::At
     # construct the fully qualified plot type, from the possible incomplete (abstract)
     # PlotType
     FinalType = Combined{Typ, ArgTyp}
-    plot_attributes, scene_attributes = merged_get!(()-> default_theme(scene, FinalType), plotsym(FinalType), scene, attributes)
-    # We allow to set scene attributes in the theme of a plot
-    # This is a bit shady, since it's a global setting affecting subsequent plots
-    # But lets stick with this for now, to make buttons/slider etc more usable
-    # TODO do this better
-    for key in (:raw, :camera)
-        if haskey(plot_attributes, key)
-            scene_attributes[key] = pop!(plot_attributes, key)
-        end
-    end
+    plot_attributes = merged_get!(
+        ()-> default_theme(scene, FinalType),
+        plotsym(FinalType), scene, attributes
+    )
+
+    # Transformation is a field of the plot type, but can be given as an attribute
     trans = get(plot_attributes, :transformation, automatic)
     transformation = if to_value(trans) == automatic
         Transformation(scene)
@@ -444,7 +442,6 @@ function (PlotType::Type{<: AbstractPlot{Typ}})(scene::SceneLike, attributes::At
         transform!(t, to_value(trans))
         t
     end
-
     replace_automatic!(plot_attributes, :model) do
         transformation.model
     end
@@ -452,7 +449,7 @@ function (PlotType::Type{<: AbstractPlot{Typ}})(scene::SceneLike, attributes::At
     plot_obj = FinalType(scene, transformation, plot_attributes, input, seperate_tuple(args))
     transformation.parent[] = plot_obj
     calculated_attributes!(plot_obj)
-    plot_obj, scene_attributes
+    plot_obj
 end
 
 
@@ -525,7 +522,6 @@ eval(default_plot_signatures(:plot, :plot!, :Any))
 plotfunc(::Combined{F}) where F = F
 
 
-
 """
 Main plotting signatures that plot/plot! route to if no Plot Type is given
 """
@@ -585,15 +581,45 @@ function show_attributes(attributes)
     end
 end
 
+"""
+    extract_scene_attributes!(attributes)
+
+removes all scene attributes from `attributes` and returns them in a new
+Attribute dict.
+"""
+function extract_scene_attributes!(attributes)
+    scene_attributes = (
+        :backgroundcolor,
+        :resolution,
+        :show_axis,
+        :show_legend,
+        :scale_plot,
+        :center,
+        :axis,
+        :axis2d,
+        :axis3d,
+        :legend,
+        :camera,
+        :limits,
+        :padding,
+        :raw
+    )
+    result = Attributes()
+    for k in scene_attributes
+        haskey(attributes, k) && (result[k] = pop!(attributes, k))
+    end
+    return result
+end
+
 function plot!(scene::SceneLike, ::Type{PlotType}, attributes::Attributes, input::NTuple{N, Node}, args::Node) where {N, PlotType <: AbstractPlot}
     # create "empty" plot type - empty meaning containing no plots, just attributes + arguments
-    plot_object, scene_attributes = PlotType(scene, attributes, input, args)
-
-    nattributes, rest = merge_attributes!(scene_attributes, theme(scene))
+    scene_attributes = extract_scene_attributes!(attributes)
+    plot_object = PlotType(scene, copy(attributes), input, args)
     # TODO warn about rest - should be unused arguments!
-    empty!(scene.attributes)
     # transfer the merged attributes from theme and user defined to the scene
-    merge!(scene.attributes, nattributes)
+    for (k, v) in scene_attributes
+        scene.attributes[k] = v
+    end
     for (at1, at2) in mutual_exclusive_attributes(PlotType)
         #nothing here to get around defaults in GLVisualize
         haskey(attributes, at1) && haskey(attributes, at2) && error("$at1 conflicts with $at2, please specify only one.")
@@ -606,52 +632,66 @@ function plot!(scene::SceneLike, ::Type{PlotType}, attributes::Attributes, input
     # call user defined recipe overload to fill the plot type
     plot!(plot_object)
 
-    push!(scene.plots, plot_object)
+    push!(scene, plot_object)
 
-    scene[:raw][] || update_limits!(scene)
-    (!scene[:raw][] || scene[:camera][] != automatic) && setup_camera!(scene)
-    scene[:raw][] || add_axis!(scene, rest)
+    if !scene.raw[] || scene[:camera][] !== automatic
+        # if no camera controls yet, setup camera
+        setup_camera!(scene)
+    end
+    if !scene.raw[]
+        add_axis!(scene, scene.attributes)
+    end
     # ! ∘ isaxis --> (x)-> !isaxis(x)
     # move axis to front, so that scene[end] gives back the last plot and not the axis!
     if !isempty(scene.plots) && isaxis(last(scene.plots))
         axis = pop!(scene.plots)
         pushfirst!(scene.plots, axis)
     end
-    # call the assembly recipe, that also adds this to the scene
-    # kw_args not consumed by PlotType will be passed forward to plot! as non_plot_kwargs
-    #plot!(scene, plot_object, scene_attributes)
     scene
 end
 
 function plot!(scene::Combined, ::Type{PlotType}, attributes::Attributes, args...) where PlotType <: AbstractPlot
     # create "empty" plot type - empty meaning containing no plots, just attributes + arguments
-    plot_object, scene_attributes = PlotType(scene, attributes, args)
+    plot_object = PlotType(scene, attributes, args)
+    # call user defined recipe overload to fill the plot type
+    plot!(plot_object)
+    push!(scene.plots, plot_object)
+    scene
+end
+function plot!(scene::Combined, ::Type{PlotType}, attributes::Attributes, input::NTuple{N,Node}, args::Node) where {N, PlotType <: AbstractPlot}
+    # create "empty" plot type - empty meaning containing no plots, just attributes + arguments
+    plot_object = PlotType(scene, attributes, input, args)
     # call user defined recipe overload to fill the plot type
     plot!(plot_object)
     push!(scene.plots, plot_object)
     scene
 end
 
+function apply_camera!(scene::Scene, cam_func)
+    if cam_func in (cam2d!, cam3d!, campixel!, cam3d_cad!)
+        cam_func(scene)
+    else
+        error("Unrecognized `camera` attribute type: $(typeof(cam_func)). Use automatic, cam2d! or cam3d!, campixel!, cam3d_cad!")
+    end
+end
 
 
 
 
 function setup_camera!(scene::Scene)
-    if scene[:camera][] == automatic
+    theme_cam = scene[:camera][]
+    if theme_cam == automatic
         cam = cameracontrols(scene)
+        # only automatically add camera when cameracontrols are empty (not set)
         if cam == EmptyCamera()
             if is2d(scene)
-                #@info("setting camera to 2D")
                 cam2d!(scene)
             else
-                #@info("setting camera to 3D")
                 cam3d!(scene)
             end
         end
-    elseif scene[:camera][] in (cam2d!, cam3d!, campixel!, cam3d_cad!)
-        scene[:camera][](scene)
     else
-        error("Unrecogniced `camera` attribute type: $(typeof(scene[:camera][])). Use automatic, cam2d! or cam3d!")
+        apply_camera!(scene, theme_cam)
     end
     scene
 end
@@ -668,20 +708,18 @@ function find_in_plots(scene::Scene, key::Symbol)
     end
 end
 
-
-
 function add_axis!(scene::Scene, attributes = Attributes())
-    show_axis = scene[:show_axis][]
+    show_axis = scene.show_axis[]
     show_axis isa Bool || error("show_axis needs to be a bool")
-    axistype = if scene[:axis_type][] == automatic
+    axistype = if scene.axis_type[] == automatic
         is2d(scene) ? axis2d! : axis3d!
-    elseif scene[:axis_type][] in (axis2d!, axis3d!)
-        scene[:axis_type][]
+    elseif scene.axis_type[] in (axis2d!, axis3d!)
+        scene.axis_type[]
     else
         error("Unrecogniced `axis_type` attribute type: $(typeof(scene[:axis_type][])). Use automatic, axis2d! or axis3d!")
     end
 
-    if show_axis && !(any(isaxis, plots(scene)))
+    if show_axis && scene[Axis] === nothing
         axis_attributes = Attributes()
         for key in (:axis, :axis2d, :axis3d)
             if haskey(scene, key) && !isempty(scene[key])
@@ -695,8 +733,12 @@ function add_axis!(scene::Scene, attributes = Attributes())
         labels = get(attributes, :ticklabels) do
             find_in_plots(scene, :ticklabels)
         end
+        lims = lift(scene.limits, scene.data_limits) do sl, dl
+            sl === automatic && return dl
+            return sl
+        end
         axistype(
-            scene, axis_attributes, limits(scene),
+            scene, axis_attributes, lims,
             ticks = (ranges = ranges, labels = labels)
         )
     end
@@ -704,69 +746,9 @@ function add_axis!(scene::Scene, attributes = Attributes())
 end
 
 function add_labels!(scene::Scene)
-    if plot_attributes[:show_legend][] && haskey(p.attributes, :colormap)
+    if plot_attributes.show_legend[] && haskey(p.attributes, :colormap)
         legend_attributes = plot_attributes[:legend][]
         colorlegend(scene, p.attributes[:colormap], p.attributes[:colorrange], legend_attributes)
     end
-    scene
-end
-
-"""
-    update_limits!(scene::Scene)
-
-This function updates the limits of the `Scene` passed to it based on its data.
-"""
-update_limits!(scene::Scene) = update_limits!(scene, scene[:limits][], scene[:padding][])
-
-function update_limits!(scene::Scene, limits::Automatic, padding)
-    # for when scene is empty
-    dlimits = data_limits(scene)
-    tlims = (minimum(dlimits), maximum(dlimits))
-    if !all(x-> all(isfinite, x), tlims)
-        @warn "limits of scene contain non finite values: $(tlims[1]) .. $(tlims[2])"
-        mini = map(x-> ifelse(isfinite(x), x, 0.0), tlims[1])
-        maxi = Vec3f0(ntuple(3) do i
-            x = tlims[2][i]
-            ifelse(isfinite(x), x, tlims[1][i] + 1f0)
-        end)
-        tlims = (mini, maxi)
-    end
-    new_widths = Vec3f0(ntuple(3) do i
-        a = tlims[1][i]; b = tlims[2][i]
-        w = b - a
-        # check for widths == 0.0... 3rd dimension is allowed to be 0 though.
-        # TODO maybe we should allow any one dimension to be 0, and then use the other 2 as 2D
-        with0 = (i != 3) && (w ≈ 0.0)
-        with0 && @warn "Founds 0 width in scene limits: $(tlims[1]) .. $(tlims[2])"
-        ifelse(with0, 1f0, w)
-    end)
-    update_limits!(scene, FRect3D(tlims[1], new_widths), padding)
-end
-
-"""
-    update_limits!(scene::Scene, new_limits::HyperRectangle, padding = Vec3f0(0))
-
-This function updates the limits of the given `Scene` according to the given HyperRectangle.
-
-A `HyperRectangle` is a generalization of a rectangle to n dimensions.  It contains two vectors.
-The first vector defines the origin; the second defines the displacement of the vertices from the origin.
-This second vector can be thought of in two dimensions as a vector of width (x-axis) and height (y-axis),
-and in three dimensions as a vector of the width (x-axis), breadth (y-axis), and height (z-axis).
-
-Such a `HyperRectangle` can be constructed using the `FRect` or `FRect3D` functions that are exported by
-`AbstractPlotting.jl`.  See their documentation for more information.
-"""
-function update_limits!(scene::Scene, new_limits::HyperRectangle, padding = Vec3f0(0))
-    lims = FRect3D(new_limits)
-    lim_w = widths(lims)
-    # use the smallest widths for scaling, to have a consistently wide padding for all sides
-    minw = if lim_w[3] ≈ 0.0
-        m = min(lim_w[1], lim_w[2])
-        Vec3f0(m, m, 0.0)
-    else
-        Vec3f0(minimum(lim_w))
-    end
-    padd_abs = minw .* to_ndim(Vec3f0, padding, 0.0)
-    limits(scene)[] = FRect3D(minimum(lims) .- padd_abs, lim_w .+  2padd_abs)
     scene
 end