+++ src/theming.jl
@@ -36,6 +36,7 @@ const minimal_default = Attributes(
     show_legend = false,
     scale_plot = true,
     center = true,
+    update_limits = true,
     axis = Attributes(),
     axis2d = Attributes(),
     axis3d = Attributes(),
@@ -47,17 +48,16 @@ const minimal_default = Attributes(
     raw = false
 )
 
-const _current_default_theme = Attributes(; minimal_default...) # make a copy. TODO overload copy?
+const _current_default_theme = copy(minimal_default)
 
 function current_default_theme(; kw_args...)
-    new_theme, rest = merge_attributes!(Attributes(kw_args), _current_default_theme)
-    merge!(new_theme, rest)
+    return merge!(Attributes(kw_args), _current_default_theme)
 end
 
 function set_theme!(new_theme::Attributes)
     empty!(_current_default_theme)
-    new_theme, rest = merge_attributes!(new_theme, minimal_default)
-    merge!(_current_default_theme, new_theme, rest)
+    new_theme = merge!(new_theme, minimal_default)
+    merge!(_current_default_theme, new_theme)
     return
 end
 function set_theme!(;kw_args...)